pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "litegraph-mcp"
version = "1.0.0"
description = "Standalone Knowledge Graph RAG Service - Graph CRUD, semantic search, question management"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "IGame Team" }
]
keywords = ["knowledge-graph", "rag", "mcp", "sqlite", "llm"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  # MCP Protocol
  "mcp[cli]>=1.0.0",
  # Web Framework
  "fastapi>=0.100.0",
  "uvicorn[standard]>=0.23.0",
  # Graph & Data
  "networkx>=3.0",
  "numpy>=1.24.0",
  # LLM & Embeddings
  "openai>=1.0.0",
  "httpx>=0.24.0",
  # Configuration
  "pyyaml>=6.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  # Search
  "rank-bm25>=0.2.2",
  "jieba>=0.42.1",
]

[project.optional-dependencies]
embeddings = [
  "sentence-transformers>=2.2.0",
  "torch>=2.0.0"
]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.21.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
igame-graph-rag = "src.server.main:main"
igame-graph-rag-http = "src.server.main:main"
igame-graph-rag-mcp = "src.mcp_server:main"
kg-build = "src.ingest.pipeline:main"

[project.urls]
Homepage = "https://github.com/igame-team/igame-graph-rag"
Documentation = "https://github.com/igame-team/igame-graph-rag#readme"
Repository = "https://github.com/igame-team/igame-graph-rag"
